+++ app/smartbrain_3000/web/_app/immutable/chunks/DTNopo7E.js
@@ -1,0 +1,1 @@
+import"./Cp3G-jpF.js";